Split executor/select.rs into focused execution modules
Split the 728-line select.rs file into 5 focused modules organized by
execution responsibility:

- mod.rs (388 lines): Main SELECT execution coordination
- grouping.rs (160 lines): GROUP BY and aggregate functions
- join.rs (163 lines): INNER and LEFT OUTER JOIN operations
- projection.rs (24 lines): Column projection and wildcard expansion
- filter.rs (2 lines): Placeholder for future WHERE clause extraction

All modules use pub(super) visibility for clean internal APIs while
maintaining the same public interface through SelectExecutor.

Changes:
- Created crates/executor/src/select/ directory structure
- Moved AggregateAccumulator and grouping logic to grouping.rs
- Extracted JOIN operations (nested_loop_join, inner, outer) to join.rs
- Moved projection logic (project_row_combined) to projection.rs
- Updated imports and fixed unused import warnings
- All tests passing (45 tests across workspace)
- No new clippy warnings

This refactoring maintains clear separation of concerns while keeping
the main execution flow understandable in mod.rs.
