+++ crates/executor/src/select/filter.rs
@@ -1,0 +1,2 @@
+// Filter module - WHERE clause evaluation is currently integrated into main execution logic
+// This module is reserved for future extraction of filtering logic
